Add projected water ripples to water shader vBlueSky
The orthographic camera follows the ripple object. If the ripples go outside the cameras viewport, and the main camera is observing said ripples, they will stretch since, I believe, they are treated as the edge of the UVs inside the orthographic viewport. It is possible to increase the view size of the orthographic camera, just makes sure to not change during runtime as the shader receives the data in awake.

Fixes #74