Add player prefab with ice trails
Testing on current player prefab revealed that the skate animations move the players legs up rather than the torso down resulting in the player floating when the capsule collider still is in contact with the ground.
New animations were made and imported where the body is moved slightly downwards on the y axis in order to have the feet be fixed to between animations.

Fixes #135